Collapse all the Scalafix rewrites into a single file.
This allows using `sbt scalafix github:typelevel/cats/v1_0_0` for
invoking the rewrite.
