+++ scalafix/input/src/main/scala/fix/v1_0_0/RemoveCartesianBuilder.scala
@@ -1,5 +1,5 @@
 /*
-rewrite = "scala:fix.to1_0_0.RemoveCartesianBuilder"
+rewrite = "scala:fix.v1_0_0.RemoveCartesianBuilder"
  */
 package fix
 package to1_0_0
