+++ scalafix/input/src/main/scala/fix/v1_0_0/RemoveUnapply.scala
@@ -1,5 +1,5 @@
 /* 
-rewrite = "scala:fix.to1_0_0.RemoveUnapply"
+rewrite = "scala:fix.v1_0_0.RemoveUnapply"
  */
 package fix
 package to1_0_0
