+++ scalafix/input/src/main/scala/fix/v1_0_0/RenameFreeSuspend.scala
@@ -1,5 +1,5 @@
 /* 
-rewrite = "scala:fix.to1_0_0.RenameFreeSuspend"
+rewrite = "scala:fix.v1_0_0.RenameFreeSuspend"
  */
 package fix
 package to1_0_0
