+++ scalafix/input/src/main/scala/fix/v1_0_0/SimplifyEitherTLift.scala
@@ -1,5 +1,5 @@
 /*
-rewrite = "scala:fix.to1_0_0.SimplifyEitherTLift"
+rewrite = "scala:fix.v1_0_0.SimplifyEitherTLift"
  */
 package fix
 package to1_0_0
