+++ scalafix/output/src/main/scala/fix/v1_0_0/RenameReducibleMethods.scala
