+++ scalafix/rewrites/src/main/scala/fix/Cats_v1_0_0.scala
@@ -1,0 +1,199 @@
+package fix
+package v1_0_0
+
+import scalafix._
+import scalafix.syntax._
+import scala.meta._
+
+object Utils {
+
+  private[fix] def rename(
+      ctx: RewriteCtx,
+      t: Term.Name,
+      renames: Map[String, String])(implicit mirror: Mirror): Patch = {
+    renames.collect {
+      case (target, rename) if t.isSymbol(target) =>
+        ctx.replaceTree(t, rename)
+    }.asPatch
+  }
+
+  implicit class TermNameOps(t: Name) {
+    def isSymbol(s: String)(implicit mirror: Mirror): Boolean =
+      t.symbolOpt.exists(_.normalized.syntax == s)
+  }
+
+  implicit class OptionTermNameOps(t: Option[Name]) {
+    def isSymbol(s: String)(implicit mirror: Mirror): Boolean =
+      t.flatMap(_.symbolOpt).exists(_.normalized.syntax == s)
+  }
+
+}
+import Utils._
+
+
+// ref: https://github.com/typelevel/cats/pull/1745
+case class RemoveCartesianBuilder(mirror: Mirror)
+    extends SemanticRewrite(mirror) {
+
+  private[this] val cartesianBuilders =
+    (1 to 22)
+      .map(arity =>
+        s"_root_.cats.syntax.CartesianBuilder#CartesianBuilder$arity.`|@|`.")
+      .toSet +
+      "_root_.cats.syntax.CartesianOps.`|@|`."
+
+  private[this] val renames: Map[String, String] =
+    (1 to 22)
+      .map { arity =>
+        Seq(
+          s"_root_.cats.syntax.CartesianBuilder#CartesianBuilder$arity.map." -> "mapN",
+          s"_root_.cats.syntax.CartesianBuilder#CartesianBuilder$arity.imap." -> "imapN",
+          s"_root_.cats.syntax.CartesianBuilder#CartesianBuilder$arity.contramap." -> "contramapN"
+        )
+      }
+      .flatten
+      .toMap
+
+  // Hackish way to work around duplicate fixes due to recursion
+  val alreadyFixedOps = collection.mutable.Set.empty[Term.Name]
+  private[this] def replaceOpWithComma(ctx: RewriteCtx, op: Term.Name): Patch =
+    if (op.symbolOpt.exists(s =>
+        cartesianBuilders.contains(s.normalized.syntax)) && !alreadyFixedOps
+        .contains(op)) {
+      alreadyFixedOps += op
+      // remove the space before |@|
+      ctx.removeToken(ctx.tokenList.prev(op.tokens.head)) +
+        // replace |@| with ,
+        ctx.replaceTree(op, ",")
+    } else {
+      Patch.empty
+    }
+
+  private[this] def removeCartesianBuilderOp(
+      ctx: RewriteCtx,
+      applyInfix: Term.ApplyInfix): Patch = {
+    applyInfix match {
+      case Term.ApplyInfix(lhs: Term.ApplyInfix, op, _, _) =>
+        removeCartesianBuilderOp(ctx, lhs) + replaceOpWithComma(ctx, op)
+      case Term.ApplyInfix(_, op, _, _) =>
+        replaceOpWithComma(ctx, op)
+    }
+  }
+
+  def rewrite(ctx: RewriteCtx): Patch = {
+    ctx.tree.collect {
+      case t: Term.ApplyInfix => removeCartesianBuilderOp(ctx, t)
+      case t: Term.Name => rename(ctx, t, renames)
+      case t @ q"import cats.syntax.cartesian._" =>
+        ctx.replaceTree(t, "import cats.syntax.apply._")
+    }.asPatch
+  }
+}
+
+
+// ref: https://github.com/typelevel/cats/pull/1583
+case class RemoveUnapply(mirror: Mirror) extends SemanticRewrite(mirror) {
+
+  private[this] val renames = Map(
+    "_root_.cats.Traverse.Ops.traverseU." -> "traverse",
+    "_root_.cats.Foldable.Ops.traverseU_." -> "traverse_",
+    "_root_.cats.Foldable.traverseU_." -> "traverse_",
+    "_root_.cats.Traverse.Ops.sequenceU." -> "sequence",
+    "_root_.cats.Foldable.Ops.sequenceU_." -> "sequence_",
+    "_root_.cats.Foldable.sequenceU_." -> "sequence_",
+    "_root_.cats.data.Func.appFuncU." -> "appFunc",
+    "_root_.cats.free.FreeT.liftTU." -> "liftT"
+  )
+
+  private[this] def importeeName(importee: Importee): Option[Name] =
+    importee match {
+      case Importee.Name(name) => Some(name)
+      case Importee.Rename(name, _) => Some(name)
+      case _ => None
+    }
+
+  private[this] def removeImportee(
+      ctx: RewriteCtx,
+      importee: Importee,
+      fixes: Map[String, String]): Patch =
+    fixes.collect {
+      case (target, _) if importeeName(importee).isSymbol(target) =>
+        ctx.removeImportee(importee)
+    }.asPatch
+
+  def rewrite(ctx: RewriteCtx): Patch = {
+    ctx.tree.collect {
+      case t: Term.Name => rename(ctx, t, renames)
+      case t: Importee.Name => removeImportee(ctx, t, renames)
+    }.asPatch
+  }
+}
+
+
+// ref: https://github.com/typelevel/cats/pull/1709
+case class RenameFreeSuspend(mirror: Mirror) extends SemanticRewrite(mirror) {
+
+  private[this] val renames = Map(
+    "_root_.cats.free.Free.suspend." -> "defer",
+    "_root_.cats.free.TrampolineFunctions.suspend." -> "defer"
+  )
+
+  def rewrite(ctx: RewriteCtx): Patch = {
+    ctx.tree.collect {
+      case t: Term.Name => rename(ctx, t, renames)
+    }.asPatch
+  }
+
+}
+
+
+// ref: https://github.com/typelevel/cats/pull/1611
+case class RenameReducibleMethods(mirror: Mirror)
+    extends SemanticRewrite(mirror) {
+
+  private[this] val renames = Map(
+    "_root_.cats.Reducible.traverse1_." -> "nonEmptyTraverse_",
+    "_root_.cats.Reducible.Ops.traverse1_." -> "nonEmptyTraverse_",
+    "_root_.cats.Reducible.intercalate1." -> "nonEmptyIntercalate",
+    "_root_.cats.Reducible.Ops.intercalate1." -> "nonEmptyIntercalate",
+    "_root_.cats.Reducible.sequence1_." -> "nonEmptySequence_",
+    "_root_.cats.Reducible.Ops.sequence1_." -> "nonEmptySequence_"
+  )
+
+  def rewrite(ctx: RewriteCtx): Patch = {
+    ctx.tree.collect {
+      case t: Term.Name => rename(ctx, t, renames)
+    }.asPatch
+  }
+
+}
+
+
+// ref: https://github.com/typelevel/cats/pull/1614
+case class SimplifyEitherTLift(mirror: Mirror) extends SemanticRewrite(mirror) {
+
+  private[this] val leftSymbol = "_root_.cats.data.EitherTFunctions.left."
+  private[this] val rightSymbol = "_root_.cats.data.EitherTFunctions.right."
+
+  private[this] def removeWithLeadingComma(ctx: RewriteCtx, t: Tree): Patch =
+    (for {
+      leadingComma <- ctx.tokenList.leading(t.tokens.head).find(_.syntax == ",")
+    } yield {
+      val leadingSpaces = ctx.tokenList.slice(leadingComma, t.tokens.head)
+      ctx.removeToken(leadingComma) +
+        leadingSpaces.map(ctx.removeToken).asPatch +
+        ctx.removeTokens(t.tokens)
+    }).asPatch
+
+  def rewrite(ctx: RewriteCtx): Patch = {
+    ctx.tree.collect {
+      case Term.ApplyType(Term.Select(_, name), Seq(f, a, b))
+          if name.isSymbol(leftSymbol) =>
+        ctx.replaceTree(name, "leftT") + removeWithLeadingComma(ctx, a)
+      case Term.ApplyType(Term.Select(_, name), Seq(f, a, b))
+          if name.isSymbol(rightSymbol) =>
+        ctx.replaceTree(name, "pure") + removeWithLeadingComma(ctx, b)
+    }.asPatch
+  }
+
+}
